# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GuimaAPI",
        "description": "API para gerenciamento de alunos, treino, dieta, aulas e avaliações físicas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/modulos/{modulo_id}/submodulos": {
            "get": {
                "tags": [
                    "Submodulos"
                ],
                "summary": "Listar todos os submodulos de um modulo específico",
                "operationId": "b54a3167efa7fca7f5b721f2b344f1fd",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de submodulos"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Submodulos"
                ],
                "summary": "Criar um novo submodulo",
                "operationId": "dadca4cdddae9dab3a3430f12ecb1647",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Treinamento"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Modulo de treinamento"
                                    },
                                    "ordem": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "img_reference": {
                                        "type": "string",
                                        "example": "null"
                                    },
                                    "modulo_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o submodulo criado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/modulos/{modulo_id}/submodulos/{submodulo_id}": {
            "get": {
                "tags": [
                    "Submodulos"
                ],
                "summary": "Mostrar um submodulo específico por id",
                "operationId": "bfb7f9b831849947c8bb33892110793d",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    },
                    {
                        "name": "submodulo_id",
                        "in": "path",
                        "description": "Id do submodulo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o submodulo específico por id"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Submodulos"
                ],
                "summary": "Deletar um submodulo específico por id",
                "operationId": "ad06f2e968878f2f99806a5e584fde82",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    },
                    {
                        "name": "sybmodulo_id",
                        "in": "path",
                        "description": "Id do submodulo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o submodulo deletado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Submodulos"
                ],
                "summary": "Criar um novo submodulo",
                "operationId": "6b0c0185917141fca715e8a9a0cce613",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    },
                    {
                        "name": "submodulo_id",
                        "in": "path",
                        "description": "Id do submodulo",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Treinamento"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Modulo de treinamento"
                                    },
                                    "ordem": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "img_reference": {
                                        "type": "string",
                                        "example": "null"
                                    },
                                    "modulo_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o submodulo criado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}